requirements.txt
asgiref==3.3.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
Django==3.2
django-extensions==3.1.3
django-oauth-toolkit==1.5.0
django-seed==0.2.2
djangorestframework==3.11.0
Faker==8.0.0
idna==2.10
jwcrypto==0.8
oauthlib==3.1.0
Pillow==8.2.0
psycopg2-binary==2.8.6
pycparser==2.20
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
urllib3==1.26.4
channels==3.0.4
djangochannelsrestframework==0.3.0
